Extract get/set env on nodes logic
